requirements.txt
torch
transformers
sentence-transformers
pinecone-client
faiss-cpu
redis
pandas
numpy
scipy
boto3
psutil
pynvml
docker
gtts
mpg123
netmiko
jinja2
fastapi
uvicorn
datasets
optuna
matplotlib
pyaudio
opencv-python
mss
websockets
python-docx
pdfplumber
pyyaml
web3
bitcoinlib
scapy
bip39
cryptography
solana
stable-baselines3
gym
scikit-learn
pytesseract
speechrecognition
undetected-chromedriver
selenium
playwright
pyautogui
googlesearch-python
torpy
beautifulsoup4
psycopg2-binary
kaggle